splited the page to components and added a check for current account and isFirst also added a regex for valid username
